move toJson into shopper request api
